requirements.txt
django
coreapi
Markdown
django-filter
django-crispy-forms
django-guardian
djangorestframework
psycopg2
pillow
